feat: adding weight on 'add item order'

+++ src/main/kotlin/SC/ProjetoSC/request/AdicionarItemPedidoRequest.kt
@@ -5,7 +5,7 @@ import sc.projetosc.dto.InformacaoBoloDTO
 data class AdicionarItemPedidoRequest(
     val idCliente: Int? = null,
     val idProduto: Int? = null,
-    var quantidade: Int? = 0,
+    var quantidade: Double? = 0.0,
     var preco: Double = 0.0,
     val listaIngredientes: List<Int>? = null,
     val informacaoBolo: InformacaoBoloDTO? = null,
