+++ src/main/kotlin/SC/ProjetoSC/response/ItemPedidoResponse.kt
@@ -6,7 +6,7 @@ import sc.projetosc.entity.InformacaoBolo
 data class ItemPedidoResponse(
     val idItemPedido: Int? = null,
     val descricao: String? = null,
-    val quantidade: Int? = null,
+    val quantidade: Double? = null,
     val precoUnitario: Double? = null,
     val informacaoBolo: InformacaoBolo? = null,
     val ingredientes: List<ItemPedidoIngredienteResponse>? = null,
